redundancy
If this is correctly gathered data (a random sample of some random
process/problem) then you should never remove samples. This is just
characteristics of your data, you have sampled two objects which have
the same representation. Efficiently, this will weight this point more,
which is a valid thing to do if the reality is like this. In short
never remove anything unless you are 100% sure that this is due to
error.